# api/openapi.yaml
openapi: 3.1.0
info:
  title: SlateDesk API
  description: |
    SlateDesk v2 HTTP API. Milestone 1 surface: local authentication
    (signed HttpOnly session cookie), users, and teams. Milestone 2 adds
    the ticket core: tickets, articles, attachments, tags, dashboard
    counters, and the realtime event stream. Milestone 3 adds the email
    channel's admin surface: mailbox configuration with live fetch/send
    tests and the Google OAuth connect flow, the instance external-URL
    setting, and outbound delivery status on articles with a retry
    operation.

    All errors are RFC 9457 problem details (`application/problem+json`).

    Access model (M2): every ticket-core operation requires an
    authenticated **agent or admin** session — customer access arrives
    with the M4 portal. The one stricter rule: `DELETE /tags/{id}` is
    **admin only**. Agents can additionally **read** users and teams
    (`GET /users`, `/users/{id}`, `/teams`, `/teams/{id}`) — the
    workspace needs them for assignee pickers, team routing, and
    requester lookups; all user/team mutations remain admin only. Each
    operation's description states its required role.

    Access model (M3): everything under `/mailboxes` and `/settings` is
    **admin only** (including the Google OAuth callback, which arrives as
    a top-level browser navigation carrying the admin's session cookie).
    `POST /articles/{id}/retry-send` is agent or admin, like the rest of
...

...
    (`/canned-replies`).

...

...

...
servers:
...
security:
  - cookieAuth: []
  - apiKey: []
...
paths:
...
  /webhooks:
    get:
      operationId: listWebhooks
      summary: List webhooks
      description: |
        All configured webhook endpoints, newest first. The signing secret
        is never included. **Admin only.**
      tags: [webhooks]
      responses:
        "200":
          description: All webhooks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createWebhook
      summary: Create a webhook
      description: |
        Registers an outbound endpoint subscribed to one or more event
        types (`ticket.created`, `ticket.updated`, `article.created`).
        Each delivery is signed `X-SlateDesk-Signature: sha256=<hex HMAC>`
        over the raw body using the endpoint's signing secret. Supply
        `secret` to set that secret, or omit it to have one generated;
        either way the effective secret is returned **once** as
        `signing_secret` so it can be configured on the receiver, and it is
        never returned again. **Admin only.**
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "201":
          description: Webhook created; signing secret returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
  /webhooks/{id}:
    parameters:
      - $ref: "#/components/parameters/WebhookID"
    patch:
      operationId: updateWebhook
      summary: Update a webhook
      description: |
        Partial update; omitted fields are left unchanged. Supplying
        `secret` rotates the signing secret; like other credential
        rotations, the new secret is **not** echoed in the response.
        **Admin only.**
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebhookRequest"
      responses:
        "200":
          description: The updated webhook (secret never included).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook
      description: |
        Removes the endpoint and its delivery log (cascade). **Admin
        only.**
      tags: [webhooks]
      responses:
        "204":
          description: Webhook deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /webhooks/{id}/test:
    post:
      operationId: testWebhook
      summary: Send a test ping to a webhook
      description: |
        Synchronously delivers a signed `ping` payload to the endpoint and
        reports the outcome. A non-2xx response from the endpoint is a
        **200 with `ok=false`** — the ping was attempted; the result is the
        payload. The ping is not recorded in the delivery log. **Admin
        only.**
      tags: [webhooks]
      parameters:
        - $ref: "#/components/parameters/WebhookID"
      responses:
        "200":
          description: Test outcome.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTestResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sd_session
      description: |
        Signed HttpOnly session cookie (SameSite=Lax) carrying the user id
        and token_version. Obtained via POST /auth/login; cleared via
        POST /auth/logout.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sd_live_...
      description: |
        Scoped machine API key presented as
        `Authorization: Bearer sd_live_…`. An **alternative** to
        `cookieAuth` on every `/api/v1` operation: because the top-level
        `security` lists the two schemes as separate requirement objects
        (an OR), either scheme alone authenticates a request and the two
        are never both required. The key is a public `sd_live_` prefix
        plus 256 bits of base62 entropy; only its SHA-256 digest is
        stored, and the request acts as the admin or agent who created
        the key, carrying that key's scopes. Scopes are `read` and
        `write` — a `read` key is refused (403) on mutating operations.
        OpenAPI cannot bind scopes to an HTTP-bearer scheme, so the scope
        gate is enforced in the handler and documented rather than
        expressed as security scopes here.
  parameters:
...
    WebhookID:
      name: id
      in: path
      required: true
      description: Webhook id.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request body or parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, invalid, or expired session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Problem:
      description: Any other error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Media type application/problem+json.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: Problem type URI; "about:blank" when unspecified.
          default: "about:blank"
        title:
          type: string
          description: Short human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Occurrence-specific explanation.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this occurrence (request path).
...
    WebhookEvent:
      type: string
      description: A subscribable outbound event type.
      enum: [ticket.created, ticket.updated, article.created]
...
    Webhook:
      type: object
      description: |
        An outbound webhook endpoint. The signing secret is never included
        in any response.
      required: [id, name, url, events, active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
        active:
          type: boolean
          description: Inactive endpoints receive no deliveries.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required: [name, url, events]
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
          format: uri
          pattern: "^https?://"
        events:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/WebhookEvent"
        active:
          type: boolean
          default: true
        secret:
          type: string
          minLength: 1
          writeOnly: true
          description: |
            HMAC-SHA256 signing secret. Write-only: encrypted at rest and
            never echoed. Omit to have a random secret generated; the
            effective secret is returned once as `signing_secret`.
    WebhookCreated:
      description: |
        A newly created webhook including its **one-time** signing secret.
      allOf:
        - $ref: "#/components/schemas/Webhook"
        - type: object
          required: [signing_secret]
          properties:
            signing_secret:
              type: string
              description: |
                The effective HMAC-SHA256 signing secret (supplied or
                generated). Returned once; configure it on the receiver to
                verify the `X-SlateDesk-Signature` header. Never returned
                again.
    UpdateWebhookRequest:
      type: object
      description: Partial update; omitted fields are left unchanged.
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
          format: uri
          pattern: "^https?://"
        events:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/WebhookEvent"
        active:
          type: boolean
        secret:
          type: string
          minLength: 1
          writeOnly: true
          description: |
            Rotate the HMAC-SHA256 signing secret. Write-only and never
            echoed — the caller supplies the new value (no generation on
            update).
    WebhookTestResult:
      type: object
      description: |
        Outcome of a synchronous webhook test ping. `ok=false` still
        arrives as HTTP 200 — the ping was attempted; this object is its
        result.
      required: [ok, response_code, detail, latency_ms]
      properties:
        ok:
          type: boolean
          description: True when the endpoint answered 2xx.
        response_code:
          type: ["integer", "null"]
          description: |
            HTTP status returned by the endpoint; null on transport error.
        detail:
          type: string
          description: |
            Human-readable outcome (the error on failure, a short summary
            otherwise).
        latency_ms:
          type: integer
          format: int64
          minimum: 0
          description: Wall-clock duration of the ping.
...